Toggle full screen on the press of alt and enter, not on the next key event
The toggle tested the key state the previous event left behind: is_key_down is only
written at the end of UI::send_key, and the test ran for every key event. So it fired
on the release of enter, on its auto repeat, or on any other key pressed while alt and
enter were both down; holding enter, then pressing alt, then pressing any key flipped
the screen.

The repeat now comes from the event itself: key_callback passes GLFW_REPEAT along and
only the press of enter toggles. The four key flags are cleared after the toggle
because the toggle recreates the window and the release of keys held across that is
lost (a first attempt which remembered the press swallowed the next toggle instead).

The desktop size the toggle and the resolution check compare against was read in main,
before glfwInit made the process DPI aware, so Windows reported it with the display
scaling applied (2560x1440 instead of 3840x2160 at 150%). It is read after glfwInit
now, and ScreenResolution includes <Windows.h> itself instead of relying on the order
of the includes of whoever includes it.

Leaving full screen is always allowed: the old width < max_width gate made alt+enter a
no-op at or above the desktop size. The desktop size, both outcomes of the toggle and
the resolution that forces full screen are logged.

+++ Util/ScreenResolution.h
@@ -19,8 +19,7 @@
 
 #include "../Configuration.h"
 
-#include <windef.h>
-#include <WinUser.h>
+#include <Windows.h>
 
 namespace ms
 {
@@ -38,8 +37,8 @@ namespace ms
 			GetWindowRect(hDesktop, &desktop);
 
 			// The top left corner will have coordinates (0, 0) and the bottom right corner will have coordinates (horizontal, vertical)
-			Configuration::get().set_max_width(desktop.right);
-			Configuration::get().set_max_height(desktop.bottom);
+			Configuration::get().set_max_width(static_cast<int16_t>(desktop.right));
+			Configuration::get().set_max_height(static_cast<int16_t>(desktop.bottom));
 
 			// Get the state of the Caps Lock key
 			int nVirtKey = GetKeyState(20);
